Add Transport Sample Into Factory Method Design Pattern fixes #12
