+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/dialog_sample/button/events/ButtonEventListener.java
@@ -1,0 +1,6 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.dialog_sample.button.events;
+
+public interface ButtonEventListener {
+    void setOnClickListener();
+    void setOnLongClickListener();
+}
