+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/dialog_sample/button/types/ButtonType.java
@@ -1,0 +1,5 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.dialog_sample.button.types;
+
+public enum ButtonType {
+    TEXT_BUTTON, OUTLINE_BUTTON
+}
