+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/dialog_sample/dialog/types/DialogType.java
@@ -1,0 +1,5 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.dialog_sample.dialog.types;
+
+public enum DialogType {
+    NORMAL,BOTTOM_SHEET
+}
