+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/message_sample/Message.kt
@@ -1,0 +1,16 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.message_sample
+
+abstract class Message constructor(
+    val type: MessageType
+) {
+
+    private fun someOperationForAllMessages() {
+
+    }
+
+    init {
+        someOperationForAllMessages()
+    }
+
+    protected abstract fun construct()
+}
