+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/message_sample/MessageType.kt
@@ -1,0 +1,5 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.message_sample
+
+enum class MessageType {
+    VOICE, TEXT, VIDEO
+}
