+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/reserve_sample/types/ReserveType.java
@@ -1,0 +1,5 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.reserve_sample.types;
+
+public enum ReserveType {
+    YACHT, LINE, CRUISE, PLACE
+}
