+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/transport_sample/logistics/LogisticsFactory.java
@@ -1,0 +1,5 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.transport_sample.logistics;
+
+public interface LogisticsFactory {
+    public Transport createTransport(String type);
+}
