+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/transport_sample/logistics/Transport.java
@@ -1,0 +1,6 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.transport_sample.logistics;
+
+public interface Transport {
+    public void load();
+    public void drive();
+}
