+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/transport_sample/logistics/road/RoadLogistics.java
@@ -1,0 +1,16 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.transport_sample.logistics.road;
+
+import ir.vg.app.designpattern.creational.factory_method_pattern.transport_sample.logistics.Transport;
+
+public class RoadLogistics implements Transport {
+
+    @Override
+    public void load() {
+
+    }
+
+    @Override
+    public void drive() {
+
+    }
+}
