+++ app/src/main/java/ir/vg/app/designpattern/creational/factory_method_pattern/transport_sample/logistics/sea/SeaLogistics.java
@@ -1,0 +1,17 @@
+package ir.vg.app.designpattern.creational.factory_method_pattern.transport_sample.logistics.sea;
+
+import ir.vg.app.designpattern.creational.factory_method_pattern.transport_sample.logistics.Transport;
+
+public class SeaLogistics implements Transport {
+
+
+    @Override
+    public void load() {
+
+    }
+
+    @Override
+    public void drive() {
+
+    }
+}
