--- app/src/main/java/ir/vg/app/designpattern/creational/factory_pattern/dialog_sample/button/events/ButtonEventListener.java
@@ -1,6 +1,0 @@
-package ir.vg.app.designpattern.creational.factory_pattern.dialog_sample.button.events;
-
-public interface ButtonEventListener {
-    void setOnClickListener();
-    void setOnLongClickListener();
-}
