--- app/src/main/java/ir/vg/app/designpattern/creational/factory_pattern/dialog_sample/dialog/types/DialogType.java
@@ -1,5 +1,0 @@
-package ir.vg.app.designpattern.creational.factory_pattern.dialog_sample.dialog.types;
-
-public enum DialogType {
-    NORMAL,BOTTOM_SHEET
-}
