--- app/src/main/java/ir/vg/app/designpattern/creational/factory_pattern/message_sample/MessageType.kt
@@ -1,5 +1,0 @@
-package ir.vg.app.designpattern.creational.factory_pattern.message_sample
-
-enum class MessageType {
-    VOICE, TEXT, VIDEO
-}
