--- app/src/main/java/ir/vg/app/designpattern/creational/factory_pattern/message_sample/types/video/VideoMessage.kt
@@ -1,13 +1,0 @@
-package ir.vg.app.designpattern.creational.factory_pattern.message_sample.types.video
-
-import ir.vg.app.designpattern.creational.factory_pattern.message_sample.Message
-import ir.vg.app.designpattern.creational.factory_pattern.message_sample.MessageType
-
-class VideoMessage : Message(MessageType.VOICE) {
-    override fun construct() {
-        println("Building Video Message")
-    }
-    init {
-        construct()
-    }
-}
