--- app/src/main/java/ir/vg/app/designpattern/creational/factory_pattern/message_sample/types/voice/VoiceMessage.kt
@@ -1,13 +1,0 @@
-package ir.vg.app.designpattern.creational.factory_pattern.message_sample.types.voice
-
-import ir.vg.app.designpattern.creational.factory_pattern.message_sample.Message
-import ir.vg.app.designpattern.creational.factory_pattern.message_sample.MessageType
-
-class VoiceMessage : Message(MessageType.VOICE) {
-    override fun construct() {
-        println("Building Voice Message")
-    }
-    init {
-        construct()
-    }
-}
