--- app/src/main/java/ir/vg/app/designpattern/creational/factory_pattern/reserve_sample/types/ReserveType.java
@@ -1,5 +1,0 @@
-package ir.vg.app.designpattern.creational.factory_pattern.reserve_sample.types;
-
-public enum ReserveType {
-    YACHT, LINE, CRUISE, PLACE
-}
