--- app/src/main/java/ir/vg/app/designpattern/creational/factory_pattern/reserve_sample/types/yacht/YachtReserve.java
@@ -1,15 +1,0 @@
-package ir.vg.app.designpattern.creational.factory_pattern.reserve_sample.types.yacht;
-
-import ir.vg.app.designpattern.creational.factory_pattern.reserve_sample.Reserve;
-import ir.vg.app.designpattern.creational.factory_pattern.reserve_sample.types.ReserveType;
-
-public class YachtReserve extends Reserve {
-    public YachtReserve() {
-        super(ReserveType.YACHT);
-        construct();
-    }
-    @Override
-    protected void construct() {
-        System.out.println("Building Yacht Reserve");
-    }
-}
